clinic_module_data/src/main/java/kss/petclinic/clinic_module_data/services/map/OwnerMapService.java
package kss.petclinic.clinic_module_data.services.map;

import kss.petclinic.clinic_module_data.model.Owner;
import kss.petclinic.clinic_module_data.model.Pet;
import kss.petclinic.clinic_module_data.services.OwnerService;
import kss.petclinic.clinic_module_data.services.PetService;
import kss.petclinic.clinic_module_data.services.PetTypeService;
import org.springframework.context.annotation.Profile;
import org.springframework.stereotype.Service;

import java.util.Arrays;
import java.util.List;
import java.util.Set;
import java.util.stream.Stream;

import static java.util.stream.Collectors.toList;

@Service
@Profile({"default", "map"})
public class OwnerMapService extends AbstractMapService<Owner, Long> implements OwnerService {

    private final PetService petService;
    private final PetTypeService petTypeService;

    public OwnerMapService(PetService petService, PetTypeService petTypeService) {
        this.petService = petService;
        this.petTypeService = petTypeService;
    }

    @Override
    public Owner findById(Long id) {
        return super.findById(id);
    }

    @Override
    public Owner save(Owner entity) {
        if (entity != null) {

            if (entity.getPets() != null) {
                entity.getPets().forEach(pet -> {
                    if (pet.getPetType() != null) {
                        if (pet.getPetType().getId() != null) {
                            pet.setPetType(petTypeService.save(pet.getPetType()));
                        }
                    } else {
                        throw new RuntimeException("Pet type is required!!");
                    }

                    if (pet.getId() == null) {
                        Pet savedPet = petService.save(pet);
                        pet.setId(savedPet.getId());
                    }
                });
            }
            return super.save(entity.getId(), entity);

        } else {
            return null;
        }

    }

    @Override
    public Set<Owner> findAll() {
        return super.findAll();
    }

    @Override
    public void deleteById(Long id) {
        super.deleteById(id);
    }

    @Override
    public void delete(Owner object) {
        super.delete(object);
    }

    @Override
    public Owner findByLastName(String lastName) {
        List<Owner> collect = super.findAll().stream()
                .filter(owner -> lastName.equalsIgnoreCase(owner.getLastName())).collect(toList());

        return collect.get(0);
    }
}
